+++ on_audio_query/android/src/main/kotlin/com/lucasjosino/on_audio_query/query/OnAlbumsQuery.kt
@@ -44,7 +44,8 @@ class OnAlbumsQuery : ViewModel() {
         // Sort: Type and Order.
         sortType = checkAlbumSortType(
             call.argument<Int>("sortType")!!,
-            call.argument<Int>("orderType")!!
+            call.argument<Int>("orderType")!!,
+            call.argument<Boolean>("ignoreCase")!!
         )
         // Check uri:
         //   * [0]: External.