+++ on_audio_query/android/src/main/kotlin/com/lucasjosino/on_audio_query/query/OnArtistsQuery.kt
@@ -41,7 +41,8 @@ class OnArtistsQuery : ViewModel() {
         // Sort: Type and Order
         sortType = checkArtistSortType(
             call.argument<Int>("sortType")!!,
-            call.argument<Int>("orderType")!!
+            call.argument<Int>("orderType")!!,
+            call.argument<Boolean>("ignoreCase")!!
         )
         // Check uri:
         //   * [0]: External.