+++ on_audio_query/android/src/main/kotlin/com/lucasjosino/on_audio_query/query/OnGenresQuery.kt
@@ -41,7 +41,8 @@ class OnGenresQuery : ViewModel() {
         // Sort: Type and Order.
         sortType = checkGenreSortType(
             call.argument<Int>("sortType")!!,
-            call.argument<Int>("orderType")!!
+            call.argument<Int>("orderType")!!,
+            call.argument<Boolean>("ignoreCase")!!
         )
         // Check uri:
         //   * [0]: External.