+++ on_audio_query/android/src/main/kotlin/com/lucasjosino/on_audio_query/types/sorttypes/PlaylistSortType.kt
@@ -2,14 +2,18 @@ package com.lucasjosino.on_audio_query.types.sorttypes
 
 import android.provider.MediaStore
 
-fun checkPlaylistSortType(sortType: Int, order: Int): String {
+fun checkPlaylistSortType(sortType: Int, order: Int, ignoreCase: Boolean): String {
     //[ASC] = Ascending Order
     //[DESC] = Descending Order
     //TODO: **Review this code later**
-    val pOrder: String = if (order == 0) " ASC" else " DESC"
+    val orderAndCase: String = if (ignoreCase) {
+        if (order == 0) " COLLATE NOCASE ASC" else " COLLATE NOCASE DESC"
+    } else {
+        if (order == 0) " ASC" else " DESC"
+    }
     return when (sortType) {
-        0 -> MediaStore.Audio.Playlists.NAME + pOrder
-        1 -> MediaStore.Audio.Playlists.DATE_ADDED + pOrder
-        else -> MediaStore.Audio.Playlists.NAME  //
+        0 -> MediaStore.Audio.Playlists.NAME + orderAndCase
+        1 -> MediaStore.Audio.Playlists.DATE_ADDED + orderAndCase
+        else -> MediaStore.Audio.Playlists.DEFAULT_SORT_ORDER + orderAndCase
     }
 }